+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Ideas/Upcoming]
 
 ## [Unreleased]
+### Added
+* WindowStatePersistBehavior to load and save window state, position and size
 
 ## [1.1.0] - 2021-08-13
 ### Added